Add player colour as a level property
This PR adds a new XML property for the player's colour. It is 0 by
default, but you can change it to any colour ID. For example, making
the player use the trinket color is `<PlayerColour>3</PlayerColour>`.

This is mostly a quality-of-life addition, as the player's colour is
always 0 unless changed by scripts. A lot of levels which use different
player colours use an intro script which both changes the player's
colour and sets their respawn colour, which works great for finished,
completed levels, but makes playtesting a little more annoying as they
will spawn in as the wrong colour. Adding a level property for the
default player colour fixes this annoyance.

Additionally, this changes the behavior of `restoreplayercolour`. This
command used to set the player's colour to 0 (ignoring the respawn
colour, so when Viridian would die, they would revert to the respawn
colour). Now, it sets both Viridian's colour AND the respawn colour to
what was present in the level file. This way, you can temporarily
change the player colour using the script commands, and then use
`restoreplayercolour` to revert back to what the player colour
normally is.

The start point colour has also changed, to show the player colour
instead of always colour 0.

Like most changes like this, a way to change this in-editor does not
yet exist, but is planned for the future.

+++ desktop_version/src/CustomLevels.cpp
@@ -404,6 +404,8 @@ void customlevelclass::reset(void)
     script.textbox_colours.clear();
     script.add_default_colours();
     map.specialroomnames.clear();
+
+    player_colour = 0;
 }
 
 const int* customlevelclass::loadlevel( int rxi, int ryi )
@@ -1414,6 +1416,11 @@ next:
                 map.specialroomnames.push_back(name);
             }
         }
+
+        if (SDL_strcmp(pKey, "PlayerColour") == 0)
+        {
+            player_colour = help.Int(pText);
+        }
     }
 
     if (mapwidth < maxwidth)
@@ -1553,6 +1560,20 @@ bool customlevelclass::save(const std::string& _path)
         // Also get rid of this one...
         tinyxml2::XMLElement* element;
         while ((element = msg->FirstChildElement("rtl")) != NULL)
+        {
+            doc.DeleteNode(element);
+        }
+    }
+
+    if (player_colour != 0)
+    {
+        xml::update_tag(msg, "PlayerColour", player_colour);
+    }
+    else
+    {
+        // Get rid of this one as well, since older levels don't have this property anyways
+        tinyxml2::XMLElement* element;
+        while ((element = msg->FirstChildElement("PlayerColour")) != NULL)
         {
             doc.DeleteNode(element);
         }
